# Faster UOp hashing